[Timer]: Precision has been improved

+++ install_pack/ev3cleti_library/public/ev3_timer.h
@@ -2,7 +2,7 @@
 #define TIMER_H
 
 void Timer_reset();
-int32_t Timer_get_time();
+float Timer_get_time();
 void Timer_close();	
 void Timer_sleep_until(int sec);
 short Timer_count_from(int sec, int* currSec);
